# docs/swagger.yaml
basePath: /api/
definitions:
  domain.Application:
    properties:
      authorID:
        type: string
      category:
        example: app_freezing
        type: string
      createdAt:
        example: "1990-01-01T00:00:00Z"
        type: string
      emailFeedBack:
        type: string
      emailReg:
        type: string
      fullName:
        type: string
      id:
        type: integer
      status:
        example: open
        type: string
      text:
        type: string
      updatedAt:
        example: "1990-01-01T00:00:00Z"
        type: string
    type: object
...
  domain.RegisterRequest:
    properties:
      confirmPassword:
        example: "123123"
        type: string
      dob:
        example: "1990-01-01T00:00:00Z"
        type: string
      email:
        example: example@example.ru
        type: string
      firstName:
        type: string
      gender:
        type: string
      lastName:
        type: string
      password:
        example: "123123"
        type: string
    type: object
...
  domain.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
    type: object
...
  handler.ApplicationIDResponse:
    properties:
      applicationID:
        type: integer
    type: object
...
  handler.IsLoggedInResponse:
    properties:
      role:
        type: string
      userID:
        type: integer
    type: object
  handler.JSONResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  handler.updateStatusRequest:
    properties:
      status:
        type: string
    type: object
  handler.updateTextRequest:
    properties:
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a VK API.
  termsOfService: http://swagger.io/terms/
  title: VK API
  version: "1.0"
paths:
  /applications:
    get:
      consumes:
      - application/json
      description: Returns paginated list of applications. Admins see all, normal
        users see only theirs.
      parameters:
      - description: Page number
        format: int32
        in: query
        name: page
        type: integer
      - description: Items per page
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Application'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: List support applications
      tags:
      - Applications
    post:
      consumes:
      - application/json
      description: Creates a new support application. Can be created by a registered
        user or a temp session.
      parameters:
      - description: Application payload
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/domain.Application'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ApplicationIDResponse'
        "400":
          description: Invalid JSON or missing fields
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Create a new support application
      tags:
      - Applications
  /applications/{id}/status:
    put:
      consumes:
      - application/json
      description: Updates the status of an existing support application by ID.
      parameters:
      - description: Application ID
        format: int32
        in: path
        name: id
        required: true
        type: integer
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/handler.updateStatusRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID or request body
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Update the status of an application
      tags:
      - Applications
  /applications/{id}/text:
    put:
      consumes:
      - application/json
      description: Updates the text of an existing support application by ID.
      parameters:
      - description: Application ID
        format: int32
        in: path
        name: id
        required: true
        type: integer
      - description: New text
        in: body
        name: text
        required: true
        schema:
          $ref: '#/definitions/handler.updateTextRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID or request body
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Update the text of an application
      tags:
      - Applications
  /auth/isloggedin:
    get:
      description: Возвращает ID пользователя, если сессия валидна
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь авторизован
          schema:
            $ref: '#/definitions/handler.IsLoggedInResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Проверить авторизацию пользователя
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Логин с email и паролем, возвращает cookie сессии
      parameters:
      - description: Данные для входа
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/logout:
    post:
      description: Удаляет cookie сессии, разлогинивает пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Выход пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с профилем и устанавливает сессию
      parameters:
      - description: Данные для регистрации
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Регистрация пользователя
      tags:
      - auth
...
swagger: "2.0"